[sai-gen] Add support to packets and packet_and_bytes counters, also make counter generation to be adaptive to its type. (#501)
## Problem

Currently, SAI API generation is not leveraging the counter type information, but blindly use whatever specified in the `name` annotation. This leads to 2 problems:

1. We could accidentally generate incorrect SAI headers, e.g. name says it is a byte counter, but in fact it is a packet counter in our BM. 
2. We could not support counters that tracks multiple things, i.e. packets and bytes.

## What we are doing in this change

This change uses the type info in the counter to generate the correct SAI attributes, also fork the counter when the single counter in BM tracks multiple things.

As the screenshot shows below, after the change, if we change the counter type of the meter_bucket_outbound, a new SAI attribute will be created for it.

![image](https://github.com/sonic-net/DASH/assets/1533278/2e635c5d-76e8-4b57-8c4b-7866a4e40b45)

This change will not change any of existing SAI header files that we generated, but can be used for future scenarios.