Fix entry action search in saiapi.cpp.j2 (#558)
The refactoring introduced in #438 only looks for a proper action attribute in SAI objects, not entries. In case of SAI entry, it iterates through all the attributes and matches the values regardless of attribute ID.

This PR aligns action search for both objects and entries, and adds action ID attributes to missing create_entry calls in tests.