[doc] Include drop counters in HA API. (#563)
Packet drops is one of the major common problem in network that needs to be monitored. This change adds the drop counters that we need to monitor in DASH pipeline.

Since the drop counters are not too complicated and also could be counted as part of the HA design, hence we are adding them as part of the HA API doc.