[dash-sai] Update SAI API template to make SAI explorer works better. (#570)
There is a change directly committed to OCP SAI to make the DASH SAI APIs working better for SAI explorer. However, this change is not updated into DASH repo, which means all future SAI API updates from DASH could override it and any new APIs will not be covered.

To help with this, this change updates the SAI API template to have a better naming.

For references, the original OCP PR is here: https://github.com/opencomputeproject/SAI/pull/1971.