[doc] Simplify MSEE failover handling in ER gateway bypass design. (#619)
The current MSEE failover handling in ER gateway bypass design is
reverse route table based, which is flexible, powerful, but a bit
complicated.

With this change, we are going to simplify the MSEE failover handling
using tunnel learning. There are 1 minor issue we could not address with
this design, but we will update the design later which is compatible
with this design and having the problem solved.