Add single_match_priority in SaiTable annotation (#671)
If at least one of the match fields is Optional, Ternary or Range,
p4runtime v1 model requires 'priority' to be set for each table entry.
But in some dash scenarios, 'priority' is not required and is not
exposed to dash sai. For these special cases, we add a
single_match_priority field in SaiTable annotation to mark it. If it is
'true', the behavior changes are:

- Not generate attribute PRIORITY in SAI object
- Not generate field priority in SAI entry struct
- Set the priority of all table entries to a default (fixed) value 1 in
dash p4 libsai

It also adds the set/get support of p4 match kind `range`.